+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## Unreleased
+
+### Fixed
+ 
+- #41: Response builder broke header value
+
+
 ## 1.2.0 - 2016-03-29
 
 ### Added